feat: edit profile, add picture

+++ services/backend/user/user/profil/views.py
@@ -1,47 +1,44 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import logout
-from .models import UserProfile
-from .forms import UserProfileForm
+from .forms import CustomProfileForm
 from authentication.models import CustomUser
+from django.contrib.auth.models import AnonymousUser
 
 @login_required
 def user_profile(request):
-    user = request.user
-    try:
-        user_profile = UserProfile.objects.get(user=user)
-    except UserProfile.DoesNotExist:
-        return redirect('home')  # redirect to the home page
-    return render(request, 'profil/user_profile.html', 
-        {'user_profile': user_profile})
+	user = request.user
+	try:
+		user_profile = CustomUser.objects.get(username=user)
+	except CustomUser.DoesNotExist:
+		return redirect('home')  # redirect to the home page
+	return render(request, 'profil/user_profile.html',
+		{'user_profile': user})
 
 def edit_profile(request):
-    if request.method == 'POST':
-        form = UserProfileForm(request.POST, request.FILES, instance=request.user.userprofile)
-        if form.is_valid():
-            form.save()
-            return redirect('user_profile')
-    else:
-        form = UserProfileForm(instance=request.user.userprofile)
-    return render(request, 'profil/edit_profile.html', {'form': form})
-
+	if request.method == 'POST':
+		form = CustomProfileForm(request.POST, request.FILES, instance=request.user)
+		if form.is_valid():
+			form.save()
+			return redirect('user_profile')
+	else:
+		form = CustomProfileForm(instance=request.user)
+	return render(request, 'profil/edit_profile.html', {'form': form})
 
 def delete_account(request):
-    if request.method == 'POST':
-        user = request.user
-        
-        # delete the user's profile
-        try:
-            profile = UserProfile.objects.get(user=user)
-            profile.delete()
-        except UserProfile.DoesNotExist:
-            pass  # profile does not exist
-        
-        # delete the user
-        user.delete()
-        
-        logout(request)
-        return redirect('home')  # redirect to the home page
-    return render(request, 'profil/delete_account.html')
+	if request.method == 'POST':
+		user = request.user
+
+		# delete the user's profile  & verify anonymous
+		if not isinstance(user, AnonymousUser):
+			try:
+				user.delete()
+			except CustomUser.DoesNotExist:
+				pass  # profile does not exist
+
+
+			logout(request)
+			return redirect('home')  # redirect to the home page
+	return render(request, 'home.html')
 
 # Create your views here.
